requirements.txt
Django==2.2.6
peewee==2.10.1
PyJWT==1.5.2
djangorestframework==3.10.3

#Fork in DRF Docs
git+git://github.com/Moons15/django-rest-framework-docs.git#egg=django-rest-framerwork-docs

#CORS
django-cors-headers==2.1.0
django-cors-middleware==1.3.1

#JWT TOKEN
djangorestframework-jwt==1.11.0
